Handle exception when source cannot be reached and provide feedback to frontend
Implement generic SourceServiceException.
Implement specific Exceptions: SourceFieldsException and SourceNotFoundException.

Give the user explicit feedback when unable to connect to the service and use a HTTP 503.
For all other source errors, use SourceServiceException with the less explicit HTTP 500.

The URI is provided as a simple way to provide which source failed.
(It may or may not be more desirable to use the source name and not the URI.)

Make sure the appropriate modals exclusively receive the core exceptions.
Set timeout of errors to 30 seconds.