Defect 01278, Item 03960: Discovery views is broken and Search tab for Discovery View creation
Refactor discovery views design, using Solr classes instead of custom generating a solr query string.
The `buildSearch()` from SolrDiscoveryService.java has been completely removed.

Provide a custom search field to be used as a Solr "q" and redesign the facets to operate as "facet" and "facet.field".
The custom search field supports ordering, but no ordering is implemented at this time.

The parameters for `findBySlug()` have been changed such that:
- Now using Pageable, which uses "page", "sort", and "size" (and paves the way for potential multiple sort-by parameters).
- Added custom "offset" parameter to maintain current position when the limit (size) is changed.
- The query is now explicitly broken up into two parameters "field" and "value", and an empty field designates "search_all".

The Search response class has been simplified to only return informative properties: field, start, and total.
The "field" property tells the front-end what the field is so that the idea of an "all_fields" can be treated as an empty string.
It may be a good idea to use an empty string in all cases for the "all_fields" instead of the string "all_fields".

Fixed typos such as `getSinlgeResult` and `previouse`.

Search parameters, such as the search value text box, are now preserved more consistently.
Use controller methods instead of directly calling `discoveryContext.executeSearch()` in the `discovery-context.html` view.

Strip default parameters from the url bar query parameters, keeping it as short as reasonably possible.

When pagination position is too large after a search update, adjust the position.