Defect 01287: Record facet fields in URL bar, making discovery results deep-linkable
Filters from the route parameters are prefixed with "f.".
The parameter "filters" is no longer used.

When assigning multiple values to a single URL parameter field in `$location.search()`, an array needs to be used.

The parameters are received in the backend as a comma separated list.
If this list is not split into an array, then the wrong structure will be sent down.
The wrong structure would be something like `texan,narrative`, which should instead be two distinct values `texan`, and `narrative`.