Item 03977: Boolean to toggle "filter" as being required on a per-core basis
When the boolean "requiresFilter" is TRUE, and the "filter" property on a Discovery View is empty, then use "*.*" for Discovery View.
Where Filter is not allowed to be "*.*", the "requiresFilter" can be set to FALSE.

This fixes structural an accessibility issues with the `sourceToCreate.readOnly` field tag (radio group label must be associated with either ARIA or a <fieldset> and <fieldset> is a badly designed HTML tag). The `sourceToCreate.readOnly` field tag must be grouped with the appropriate class `form-group` for proper display an positioning for subsequent fields.