First row for FacetFields, SearchFields, and ResultMetadataFields are incorrect (and other fixes)
Redesign how the first row for FacetFields, SearchFields, and ResultMetadataFields are used by implementing ng-ifs and providing defaults.

Correctly calculate tabs length and make sure the active tab is always valid.
The `$scope.active` was renamed to `$scope.tabs.active` so that the length of tabs could be grouped with the active tab and make calculations consistent.
The number of tabs is 4 and not 3, so set length to 4.

Remove `populateSources()`, it only needs to be done on creation of the DiscoveryView.
Creating a DiscoveryView now directly handles initializing a default source if possible and subsequently the loading fields.

Implement MetadataField and FacetField models in the UI.

Don't add the default searchField in the scaffold.
Instead, append the default after the scaffold is used so that a new SearchField model can be explicitly created.

The "FacetFields" class is actually a single facet field and should be called "FacetField".