Item 03916: As a curator, I would like a warning about unavailable sources/destinations (collections/cores) when adding them to new objects
Adds and implements `/source/solr/test/ping` endpoint.
There are also some existing endpoints, but these are currently outdated and do more than is necessary.
Furthermore, the other endpoints produce errors and the "test/ping" endpoint is intended to produce warnings.

The ping (and only a ping) is performed when selecting any given `source` on the first tab in the Discovery View management's first tab.
If the ping fails, a warning is presented to the user.
If the user continues to the next tab despite the warning, an error will be presented as usual when trying to fetch the fields.

The errors should also be presented when the `filter` changes and not just the `source`.