Update tests
Be consistent with changes and update them in all places.
Most notable is that of using `$provide` for mocked models.
These mocked models need to be injected into models, repos, etc.. and the injection done in a controller is not available everywhere.
Creating a Mocked* type then using `$provide` to provide the mocked model solves this problem.