Work in Progress: testing working around maven+node problems and do coveralls tweaks.
The maven + node code, specifically exec-npm-link, attempts to make changes to the system-level (aka global) node path.
This access is prohibited on github's githubaction system.

Attempt to use the existing node processing code while under 'java' in hopes that this somehow fixes the situation.

Also do some testing with coveralls setup and configuration.
Some of the data was left out in a previous commit.