Work in Progress: test without restore-keys
The idea here is that by appending "-test", it can be determined what the default behavior for restore-keys is when undefined.
This will helps us avoid assuming the behavior.

If the default behavior does find the nearest match, then the nearest match should be the existing hash (without the "-test").