Issue 401: Cloning modals unresponsive after validation error.
The problem is that the id is removed from the form before submitting.
This is done to trigger the generation of a new ID automatically on the backend.
If the validation fails, the form no longer has the ID.

Fix this by saving the ID and restoring the ID to the form if clone submit did not succeed.