Issue 476: Sort order toggle button appends value=undefined to URL, breaking the search.
There are several areas where the is undefined logic can be improved.
Fixing these does not solve the problem and such changes have been omitted.

The problem is that the form has an initial value that is undefined.
When the form gets submitted, via a search button click or a sort button click, then the undefined gets translated into the string 'undefined'.

The immediate solution is to always initialize the value as an empty string before processing any of the URL arguments on first load.