Enable support for multiple SOLR cores via Docker build arguments.
Adding additional cores for development and testing should be easier.
The goal here is to make this easier by utilizing docker arguments.

The solr user is also abstracted into an argument for possible arg-based changes, if ever.

The default core is abstracted into MAIN_CORE docker build argument, with a default of `sage-discovery`.
The additional cores are abstracted into `EXTRA_CORES` docker build argument.
This is done primarily because docker has very poor (non-existent) loop and conditional handling.
The `COPY` command needs to be performed in a loop or in conditions but such support is unavailable.

This takes the approach of checking if `EXTRA_CORES` is empty and if not loop over space separated core names.
The `MAIN_CORE` must be pre-created so that the configuration can be copied.
The copying of the configuration in this way avoids needing to use `COPY` and instead uses the system `cp` command.

The loops can continue on even in error so `&&`, `||`, and `exit 1` are used to guarantee that the loop fails on any error.

The two chmod commands are merged into a single command.

Example docker build usage:
```
docker build --build-arg=EXTRA_CORES="extra1 extra2" -t local_solr .
```