Issue 506: Make database pre-populate specific to on 'create'.
Use `spring.sql.init.platform` to designate the file to use.
It seems that we have been settings this for no apparent reason in the past.
Change the `platform` options to `create`, `update`, and `none.

I chose `update` if we ever need to have a `data-update.sql` that gets run on every update.
(For example, a hypothetical database property that has 'last_started' database table and property that gets set to the date on every system start.)

The documentation specifically states the following:
> This allows you to switch to database-specific scripts if necessary. For example, you might choose to set it to the vendor name of the database (hsqldb, h2, oracle, mysql, postgresql, and so on).

This tells me that the previous choice of using something like `postgresql` or `h2` is superficial given the lack of utilization.
If we want databae-specific versions, we could have additional platforms as needed.
All we need to do to disable this is to set this to something like `none`.

This change has a downside of requiring yet another step in the deployment process.

All systems that use environment variables to designate properties should now be sure to set `spring.sql.init.platform` to `update`.

see: https://docs.spring.io/spring-boot/docs/current/reference/html/howto.html#howto.data-initialization.using-basic-sql-scripts
see: https://docs.spring.io/spring-boot/docs/2.1.x/reference/html/howto-database-initialization.html#howto-initialize-a-database-using-spring-jdbc