Issue 494: Provide more granular links that point to the currently selecting searching algorithms.
We are using the Github Wiki which lacks most of the structural functionality that would allow for the code here to be simpler.
Ideally, something like "id" tags can be used on a single page to only need a single link.
This is not possible with the current Github Wiki functionality.

This solution, which is far from ideal, uses a separate URL for each of the combinations between the Parser and the Operands.

Text with a link is provided on the management page because landscape is readily available.