Default Field and wild cards don't work as expected.
One would expect that when the default field is set to say `origin`, then the default wild card would be against the `origin`.
This is not the case.

If there is no value but there is a search field, then prepend the search field before the wild card like `origin:*`.