Enable support for configsets in Solr Docker image.
The `core.properties` is the default main one.
A `legacy.properties` is provided as an example and as a convenience for using the older schema.
All additional cores will use the default unless a properties file is named after it.

This also adds the ability to customize the solr configuration, such as changing port and access control.
This is done by copying the default solr.xml.
Make any modifications to this file for those changes to be copied over.

This also fixes a security concern where the core directories and files are given world/global read and write access.
See the `chmod` commands near the bottom of the `Dockerfile`.

Example:
```shell
docker build --build-arg=MAIN_CORE="sage-discovery-1" --build-arg=EXTRA_CORES="sage-discovery-2 sage-discovery-3 sage-discovery-4 legacy" -t local_solr .
```

That example will build 5 cores, the first 4 will use the default `core.properties`.
The last core, named `legacy`, will use the `legacy.core`.