Regression in boldMatch filter causing autocomplete drop down to not work as expected.
The autocomplete drop down is requiring the input variable to be modified.
The return value is not being processed at all.

When there is no matches, the input value must still be modified, even if the modification makes it empty.

While investigating this, a ton of white space ended up being auto-corrected.
I decided to leave those corrections.