Update Sage Solr to work with Solr 9.3 or greater.
The solr Dockerfile is not needed anymore, instead just use the upstream pre-built docker image.
The Dockerfile can be found here:
- https://github.com/apache/solr-docker/blob/7e3be2fce98be1fe9c80be0727b4600d4b52fd0f/9.3-slim/Dockerfile

Be more consistent with the 9.3 styling and switch:
- `text` to `text_general`.
- `dates` to `pdates`.
- `bools` back to `booleans`.
- `longs` to `plongs`.
- `doubles` to `pdoubles`.

Explicitly add `multiValued="true"` because the default when not specified is no longer multi-valued.

The `managed-schema` file must now have the XML file extension: `managed-schema.xml`.

Refresh the `solrconfig.xml` to use the one from Solr 9.3.

Update readme documentation to describe how to build the solr using 9.3 or greater.