Do not restrict fields filter to fields that have data types.
In practice, restricting the fields to only ones that have data makes the UI hard to use.
Trying to configure a Discovery View is causing problems with being unable to select the data.
When UI caching is mixed in, some times the fields show up and other times they do not.

This change makes the field display more consistently and allows for all fields to be seen and selected from.