Use "string" types for fields and supply copy fields as appropriate.
The "text_general" (which is why the dynamic field "*_text" uses) breaks a string apart.
This does not make sense for the several of these non-core optional fields.

Use Field types that are similar to the legacy version and provide copy fields that seem appropriate for use cases where "_text" or "_string" is needed.
Some of these are clearly URLs or similar and so do not provide a "_text" for these.

The "_facet" copy fields should be added as well.